Move `_exponentiate_commuting`
